+++ src/pages/Inquiry/Inquire/components/InquireForm/RequiredLabel/RequiredLabel.module.scss
