+++ src/pages/Inquiry/Inquire/components/InquireForm/RequiredLabel/index.tsx
