+++ src/pages/Inquiry/Inquiry/components/InquiryList/InquiryBlock/InquiryBlock.module.scss
